create `global_context` module
