+++ Cargo.toml
@@ -92,11 +92,13 @@ re_video = { path = "crates/utils/re_video", version = "=0.23.0-alpha.1", defaul
 re_blueprint_tree = { path = "crates/viewer/re_blueprint_tree", version = "=0.23.0-alpha.1", default-features = false }
 re_component_ui = { path = "crates/viewer/re_component_ui", version = "=0.23.0-alpha.1", default-features = false }
 re_context_menu = { path = "crates/viewer/re_context_menu", version = "=0.23.0-alpha.1", default-features = false }
-re_data_ui = { path = "crates/viewer/re_data_ui", version = "=0.23.0-alpha.1", default-features = false }
 re_chunk_store_ui = { path = "crates/viewer/re_chunk_store_ui", version = "=0.23.0-alpha.1", default-features = false }
+re_data_ui = { path = "crates/viewer/re_data_ui", version = "=0.23.0-alpha.1", default-features = false }
 re_renderer = { path = "crates/viewer/re_renderer", version = "=0.23.0-alpha.1", default-features = false }
 re_renderer_examples = { path = "crates/viewer/re_renderer_examples", version = "=0.23.0-alpha.1", default-features = false }
 re_selection_panel = { path = "crates/viewer/re_selection_panel", version = "=0.23.0-alpha.1", default-features = false }
+re_time_panel = { path = "crates/viewer/re_time_panel", version = "=0.23.0-alpha.1", default-features = false }
+re_ui = { path = "crates/viewer/re_ui", version = "=0.23.0-alpha.1", default-features = false }
 re_view = { path = "crates/viewer/re_view", version = "=0.23.0-alpha.1", default-features = false }
 re_view_bar_chart = { path = "crates/viewer/re_view_bar_chart", version = "=0.23.0-alpha.1", default-features = false }
 re_view_spatial = { path = "crates/viewer/re_view_spatial", version = "=0.23.0-alpha.1", default-features = false }
@@ -107,8 +109,6 @@ re_view_tensor = { path = "crates/viewer/re_view_tensor", version = "=0.23.0-alp
 re_view_text_document = { path = "crates/viewer/re_view_text_document", version = "=0.23.0-alpha.1", default-features = false }
 re_view_text_log = { path = "crates/viewer/re_view_text_log", version = "=0.23.0-alpha.1", default-features = false }
 re_view_time_series = { path = "crates/viewer/re_view_time_series", version = "=0.23.0-alpha.1", default-features = false }
-re_time_panel = { path = "crates/viewer/re_time_panel", version = "=0.23.0-alpha.1", default-features = false }
-re_ui = { path = "crates/viewer/re_ui", version = "=0.23.0-alpha.1", default-features = false }
 re_viewer = { path = "crates/viewer/re_viewer", version = "=0.23.0-alpha.1", default-features = false }
 re_viewer_context = { path = "crates/viewer/re_viewer_context", version = "=0.23.0-alpha.1", default-features = false }
 re_viewport = { path = "crates/viewer/re_viewport", version = "=0.23.0-alpha.1", default-features = false }
