+++ crates/viewer/re_viewer_context/src/global_context/command_sender.rs
