+++ crates/viewer/re_viewer_context/src/global_context/component_ui_registry.rs
