+++ crates/viewer/re_viewer_context/src/viewer_context.rs
@@ -1,9 +1,3 @@
-//! This module introduces a hierarchy of contexts from general to increasingly
-//! specific. Each part of the viewer should only have access a narrow context.
-//! The hierarchy looks as follows:
-//!
-//! [`ViewerContext`] -> [`GlobalContext`]
-
 use ahash::HashMap;
 use arrow::array::ArrayRef;
 use parking_lot::RwLock;
@@ -13,6 +7,7 @@ use re_entity_db::entity_db::EntityDb;
 use re_query::StorageEngineReadGuard;
 
 use crate::drag_and_drop::DragAndDropPayload;
+use crate::GlobalContext;
 use crate::{
     query_context::DataQueryResult, AppOptions, ApplicationSelectionState, CommandSender,
     ComponentUiRegistry, DragAndDropManager, IndicatedEntities, ItemCollection,
@@ -61,38 +56,6 @@ impl ViewerContext<'_> {
     pub fn command_sender(&self) -> &CommandSender {
         self.global_context.command_sender
     }
-}
-
-// TODO: move this to correct place
-// TODO: reorder to match Antoine's PR
-/// Application context that is shared across all parts of the viewer.
-pub struct GlobalContext<'a> {
-    /// Global options for the whole viewer.
-    pub app_options: &'a AppOptions,
-
-    /// Runtime info about components and archetypes.
-    ///
-    /// The component placeholder values for components are to be used when [`crate::ComponentFallbackProvider::try_provide_fallback`]
-    /// is not able to provide a value.
-    ///
-    /// ⚠️ In almost all cases you should not use this directly, but instead use the currently best fitting
-    /// [`crate::ComponentFallbackProvider`] and call [`crate::ComponentFallbackProvider::fallback_for`] instead.
-    pub reflection: &'a re_types_core::reflection::Reflection,
-
-    /// How to display components.
-    pub component_ui_registry: &'a ComponentUiRegistry,
-
-    /// Registry of all known classes of views.
-    pub view_class_registry: &'a ViewClassRegistry,
-
-    /// The [`egui::Context`].
-    pub egui_ctx: &'a egui::Context,
-
-    /// The global `re_renderer` context, holds on to all GPU resources.
-    pub render_ctx: &'a re_renderer::RenderContext,
-
-    /// Interface for sending commands back to the app
-    pub command_sender: &'a CommandSender,
 }
 
 /// Common things needed by many parts of the viewer.
